Move AccessControlList cache to own module
