+++ src/ggrc/models/hooks/acl/cache.py
@@ -1,0 +1,27 @@
+# Copyright (C) 2018 Google Inc.
+# Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+
+"""Basic cache used by program & audit roles hooks"""
+
+from ggrc.models import all_models
+
+
+class AccessControlListCache(object):  # pylint: disable=R0903
+  """Access Control List helper"""
+  def __init__(self):
+    self.cache = set()
+
+  def add(self, obj, parent, person, role_id):
+    """Add new item if it wasn't already added"""
+    key = (obj.id, obj.type, person.id, role_id, parent.id)
+    if key in self.cache:
+      return
+    self.cache.add(key)
+    acl = all_models.AccessControlList(
+        object_id=obj.id,
+        object_type=obj.type,
+        parent=parent,
+        person=person,
+        ac_role_id=role_id)
+    if hasattr(obj, "access_control_list"):
+      obj.access_control_list.append(acl)
